+++ core/services/keystore/eth.go
@@ -23,6 +23,7 @@ type Eth interface {
 	GetAll() ([]ethkey.KeyV2, error)
 	Create(chainID *big.Int) (ethkey.KeyV2, error)
 	Add(key ethkey.KeyV2, chainID *big.Int) error
+	Update(id string, maxGasGwei *big.Int) (ethkey.KeyV2, error)
 	Delete(id string) (ethkey.KeyV2, error)
 	Import(keyJSON []byte, password string, chainID *big.Int) (ethkey.KeyV2, error)
 	Export(id string, password string) ([]byte, error)
@@ -197,6 +198,28 @@ func (ks *eth) Export(id string, password string) ([]byte, error) {
 		return nil, err
 	}
 	return key.ToEncryptedJSON(password, ks.scryptParams)
+}
+
+func (ks *eth) Update(id string, maxGasGwei *big.Int) (ethkey.KeyV2, error) {
+	ks.lock.Lock()
+	defer ks.lock.Unlock()
+	if ks.isLocked() {
+		return ethkey.KeyV2{}, ErrLocked
+	}
+	key, err := ks.getByID(id)
+	if err != nil {
+		return ethkey.KeyV2{}, err
+	}
+	keyState := ks.keyStates.Eth[key.ID()]
+	keyState.MaxGasGwei = *utils.NewBig(maxGasGwei)
+	err = ks.save(func(db *gorm.DB) error {
+		return db.Model(keyState).Update("max_gas_gwei", keyState.MaxGasGwei).Error
+	})
+	if err != nil {
+		return ethkey.KeyV2{}, errors.Wrap(err, "unable to update eth key")
+	}
+	ks.notify()
+	return key, nil
 }
 
 func (ks *eth) Delete(id string) (ethkey.KeyV2, error) {