+++ core/store/migrate/migrations/0079_eth_keys_max_gas_price.sql
@@ -1,0 +1,3 @@
+-- +goose Up
+ALTER TABLE eth_key_states ADD COLUMN max_gas_gwei numeric(78,0);
+-- +goose Down